Switched type to Long

+++ src/test/java/krabban91/kodvent/kodvent/y2019/d22/DealingActionTest.java
@@ -6,6 +6,8 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import java.util.stream.LongStream;
+import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -14,7 +16,7 @@ public class DealingActionTest {
 
     @Test
     public void handle() {
-        List<Integer> input = IntStream.range(0, 10).boxed().collect(Collectors.toList());
+        List<Long> input = LongStream.range(0, 10).boxed().collect(Collectors.toList());
         assertThat(new DealingAction("deal with increment 3").handle(input)).isEqualTo(Arrays.asList(0, 7, 4, 1, 8, 5, 2, 9, 6, 3));
         assertThat(new DealingAction("cut -4").handle(input)).isEqualTo(Arrays.asList(6, 7, 8, 9, 0, 1, 2, 3, 4, 5));
         assertThat(new DealingAction("cut 3").handle(input)).isEqualTo(Arrays.asList(3, 4, 5, 6, 7, 8, 9, 0, 1, 2));
